Move computeCostModifier to the Good subclasses
Changes:
- Remove the computeCostModifier from Good to specific subclasses
- Removed the redundant proxy methods from GoodUtil as Goods supports
  the logic
- Updated header on new classes

Relates to #598
