+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/goods/EquipmentGood.java
@@ -1,13 +1,25 @@
+/*
+ * Mars Simulation Project
+ * EquipmentGood.java
+ * @date 2022-06-26
+ * @author Barry Evans
+ */
 package org.mars_sim.msp.core.goods;
 
 import org.mars_sim.msp.core.equipment.EquipmentFactory;
 import org.mars_sim.msp.core.equipment.EquipmentType;
 
-class EquipmentGood extends Good {
+/**
+ * THis represents the attributes of how an Equipment can be traded
+ */
+public class EquipmentGood extends Good {
+
+    private static final double EVA_SUIT_VALUE = 50;
+	private static final double CONTAINER_VALUE = .1;
 
     private EquipmentType equipmentType;
 
-    public EquipmentGood(EquipmentType type) {
+    EquipmentGood(EquipmentType type) {
         super( type.getName(), EquipmentType.getResourceID(type));
         this.equipmentType = type;
     }
@@ -33,5 +45,13 @@ class EquipmentGood extends Good {
         }
         
         return GoodType.CONTAINER;
+    }
+
+    @Override
+    protected double computeCostModifier() {
+        if (equipmentType == EquipmentType.EVA_SUIT) {
+			return EVA_SUIT_VALUE;
+		}
+        return CONTAINER_VALUE;
     }
 }
