+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/goods/VehicleGood.java
@@ -1,20 +1,34 @@
+/*
+ * Mars Simulation Project
+ * VehicleGood.java
+ * @date 2022-06-26
+ * @author Barry Evans
+ */
 package org.mars_sim.msp.core.goods;
 
 import org.mars_sim.msp.core.SimulationConfig;
 import org.mars_sim.msp.core.vehicle.VehicleConfig;
 import org.mars_sim.msp.core.vehicle.VehicleType;
 
+/**
+ * This represents a type of Vehicle that can be traded.
+ */
 class VehicleGood extends Good {
+	
+	private static final int VEHICLE_VALUE = 20;
+	private static final int LUV_VALUE = 750;
+	private static final int DRONE_VALUE = 50;
 
     // TODO, should load of an instance and not a static
     private static VehicleConfig vehicleConfig = SimulationConfig.instance().getVehicleConfiguration();
 
     private GoodType goodType;
+    private VehicleType vehicleType;
 
     public VehicleGood(String name) {
         super(name, VehicleType.convertName2ID(name));
 
-        VehicleType vehicleType = VehicleType.convertNameToVehicleType(name);
+        this.vehicleType = VehicleType.convertNameToVehicleType(name);
         switch(vehicleType) {
 		case DELIVERY_DRONE:
 		case LUV:
@@ -52,5 +66,17 @@ class VehicleGood extends Good {
     @Override
     public GoodType getGoodType() {
         return goodType;
+    }
+
+    @Override
+    protected double computeCostModifier() {
+        switch(vehicleType) {
+            case LUV:
+                return LUV_VALUE;
+            case DELIVERY_DRONE:
+                return DRONE_VALUE;
+            default:
+                return VEHICLE_VALUE;
+        }
     }
 }
